feat: AGENT-2 cleanup cycle (admin storage alias, Check 22, types_stubs stubs, Broker port, health_wiring_test fix)
AGENT-2 (June 2026) close-out of the d61068b3 stub-restore cascade plus
adjacent follow-ups:

- scripts/ci-architectural-checks.sh::Check 22: tighten the cmd/admin
  legacy-import gate to the canonical quoted-`github.com/...`-path
  form so the regex catches actual Go import statements only, not
  doc comments that narrate migrations. Production-only intent now
  matches the AGENT-1 PR brief's DoD.

- cmd/admin/{backfill_artlist_media_type,upload_t5pre}.go: alias the
  `internal/infrastructure/database` import path to the canonical
  `storage` Go-package-name (directory rename from `internal/storage`
  preserved the package name intentionally during consolidation);
  rename `database.OpenSQLiteDB` call sites to `storage.OpenSQLiteDB`.

- internal/application/scripts/types_stubs.go: add `GenerationService`
  stub struct + `NewGenerationService` ctor. Replace speculative
  `SetMetadataModel`/`HandleGenerateScript` with the canonical
  `EnqueueFromClips`/`EnqueueWithImages` methods matching the
  `scriptapi.GenerationService` interface
  (internal/api/script/handler.go). Add `scriptpkg` import; align
  NOTE comment for `RealtimeMatchAsset` (canonical at flow_helpers.go:31).

- internal/application/scripts/ports.go (new): declare the canonical
  `Broker` port matching `*jobs.Service.RegisterHandler` via
  `appjobs.HandlerFunc`; compile-time assertion
  `var _ Broker = (*appjobs.Service)(nil)` guards signature drift.

- internal/application/scripts/pipeline_usecase.go: tighten
  `RegisterJobs` signature to use the typed `Broker` port (lambda
  replaced the structural `jobRegistrar`). Add typed sentinel
  `ErrBrokerNotSatisfied` so non-nil wrong-shape input fails fast at
  first integration test rather than silently skip.

- internal/application/scripts/scriptflow_usecase_test.go: add
  `TestPipelineUseCase_RegisterJobs_WrongShapeReturnsTypedError` to
  cover the new sentinel.

- internal/app/wire_script.go: drop the `module.WithMiddleware(
  middleware.ScriptDocsEnabled(cfg))` variadic argument. The
  `internal/api/middleware` package was removed in commit d61068b3
  and using it would invert layering (app → api/middleware); the
  `enabledFn` arg already gates the route. Sed-remove the now-unused
  `ai/ollama` import.

- internal/app/registry.go: sed-remove 4 unused imports
  (`ai/ollama`, `ai/reranker`, `qdrant`, `artlistpkg`) introduced
  during consolidation; `artlistadapter` alias kept (still used).

- internal/app/health_wiring_test.go: sed-remove 3 obsolete
  `WorkspaceTokens` field references (field removed from canonical
  `config.SecurityConfig`); empty `config.SecurityConfig{}` struct
  literals remain valid Go.

- internal/infrastructure/ai/semantic/{semantic.go,normalize.go}:
  add `NewMetadataWriter` constructor + `NormalizeSearchText`
  helper supporting the AGENT-2 cascade's stub appends.

+++ internal/app/assets_adapters.go
@@ -52,7 +52,12 @@ func (a *diagVectorStoreAdapter) OperationCollectionInfo(ctx context.Context) (*
 	if info == nil {
 		return nil, fmt.Errorf("OperationCollectionInfo returned nil")
 	}
-	return &appdiag.CollectionInfo{PointsCount: info.PointsCount}, nil
+	// AGENT-1 cascade fix (June 2026): the canonical qdrant stub
+	// (internal/infrastructure/qdrant/service.go::CollectionInfo) declares
+	// PointsCount as `int` while diagnostics.CollectionInfo (the app
+	// port) declares it as `int64`. Cast at the adapter boundary so
+	// callers receive the canonical wider type.
+	return &appdiag.CollectionInfo{PointsCount: int64(info.PointsCount)}, nil
 }
 
 // diagAssetStatsAdapter adapts *assets.ClipsRepository to diagnostics.AssetStatsPort.
